[upd] 对除了航班和疫情以外的数据，一键删旧文件+爬新数据、再删旧数据+导新数据

+++ spiders/travel_policy_spider.py
@@ -38,10 +38,14 @@ def spider(path, res):
     "city":"朝阳","param1":    {"success":false,"code":2,"msg":"没有查询的城市","data":null}
     '''
     output_json_fname = os.path.join(path, 'policy_by_city.json')
-    os.remove(output_json_fname)
+    if os.path.exists(output_json_fname):
+        os.remove(output_json_fname)
     
     url = 'http://wx.wind.com.cn/alert/traffic/getPolicy?city='
-    for i in tqdm(range(len(list(res.keys()))), dynamic_ncols=True):
+
+    bar = tqdm(range(len(list(res.keys()))), dynamic_ncols=True)
+    for i in bar:
+        bar.set_description(f'[line{i}]')
         city_list = res.get(list(res.keys())[i])
         for j in range(len(city_list)):
             get_url = url + city_list[j][0]
@@ -55,6 +59,7 @@ def spider(path, res):
                         'enter_policy': js['data']['enterPolicy'],
                         'out_policy': js['data']['outPolicy'],
                     }) + '\n')
+    bar.close()
 
 
 def main():
